requirements.txt
anthropic==0.86.0
fastapi==0.135.2
h3==4.4.2
psycopg2==2.9.10
psycopg2_binary==2.9.9
pydantic==2.12.5
python-dotenv==1.2.2
redis==7.4.0
Requests==2.33.0
sseclient==0.0.27
uvicorn==0.42.0
